supress warnings

+++ examples/Club Penguin Revived/src/main.cpp
@@ -63,9 +63,9 @@ public:
                 glfwGetCursorPos(GJGO::g_appInstancePtr->windowPtr, &mousePosition[0], &mousePosition[1]);
                 mousePosition[1] = GJGO::Window::getHeight() - mousePosition[1];
 
-                int duration = std::abs(this->playerPosition.x - mousePosition[0]) + std::abs(this->playerPosition.y - mousePosition[1]);
+                int duration = std::abs(this->playerPosition.x - static_cast<int>(mousePosition[0])) + std::abs(this->playerPosition.y - static_cast<int>(mousePosition[1]));
 
-                this->animation = GJGO::AnimationPosition2D(duration, this->playerPosition, GJGO::Position2D{mousePosition[0], mousePosition[1]});
+                this->animation = GJGO::AnimationPosition2D(duration, this->playerPosition, GJGO::Position2D{static_cast<int>(mousePosition[0]), static_cast<int>(mousePosition[1])});
 
                 break;
             }
